Move all functions related to validating config into config_validation.py
